+++ src/main/java/de/mari_023/ae2wtlib/AE2wtlibEvents.java
@@ -2,17 +2,19 @@ package de.mari_023.ae2wtlib;
 
 import java.util.function.Consumer;
 
+import net.minecraft.network.protocol.game.ClientboundSetCarriedItemPacket;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 
+import de.mari_023.ae2wtlib.networking.ClientNetworkManager;
 import de.mari_023.ae2wtlib.networking.ServerNetworkManager;
+import de.mari_023.ae2wtlib.networking.c2s.PickBlockPacket;
 import de.mari_023.ae2wtlib.networking.s2c.UpdateRestockPacket;
 import de.mari_023.ae2wtlib.terminal.ItemWT;
 import de.mari_023.ae2wtlib.wct.CraftingTerminalHandler;
 import de.mari_023.ae2wtlib.wct.magnet_card.MagnetHandler;
 import de.mari_023.ae2wtlib.wct.magnet_card.MagnetHost;
-import de.mari_023.ae2wtlib.wct.magnet_card.MagnetMode;
 
 import appeng.api.config.Actionable;
 import appeng.api.stacks.AEItemKey;
@@ -33,7 +35,7 @@ public class AE2wtlibEvents {
         if (item.isEmpty())
             return;
         CraftingTerminalHandler cTHandler = CraftingTerminalHandler.getCraftingTerminalHandler(player);
-        if (!cTHandler.inRange() || !ItemWT.getBoolean(cTHandler.getCraftingTerminal(), "restock")
+        if (!cTHandler.inRange() || !ItemWT.getBoolean(cTHandler.getCraftingTerminal(), AE2wtlibTags.RESTOCK)
                 || cTHandler.getTargetGrid() == null || cTHandler.getTargetGrid().getStorageService() == null)
             return;
         int toAdd = Math.max(item.getMaxStackSize() / 2, 1) - count;
@@ -70,7 +72,7 @@ public class AE2wtlibEvents {
         CraftingTerminalHandler cTHandler = CraftingTerminalHandler.getCraftingTerminalHandler(player);
         ItemStack terminal = cTHandler.getCraftingTerminal();
 
-        if (!(MagnetHandler.getMagnetSettings(terminal).magnetMode == MagnetMode.PICKUP_ME))
+        if (!(MagnetHandler.getMagnetMode(terminal).pickupToME()))
             return false;
         if (!cTHandler.inRange())
             return false;
@@ -92,5 +94,54 @@ public class AE2wtlibEvents {
 
         stack.setCount(leftover);
         return leftover == 0;
+    }
+
+    public static void pickBlock(ItemStack stack) {
+        ClientNetworkManager.sendToServer(new PickBlockPacket(stack));
+    }
+
+    public static void pickBlock(ServerPlayer player, ItemStack stack) {
+        var cTHandler = CraftingTerminalHandler.getCraftingTerminalHandler(player);
+        ItemStack terminal = cTHandler.getCraftingTerminal();
+        if (!ItemWT.getBoolean(terminal, AE2wtlibTags.PICK_BLOCK))
+            return;
+
+        if (cTHandler.getTargetGrid() == null)
+            return;
+        if (cTHandler.getTargetGrid().getStorageService() == null)
+            return;
+        var networkInventory = cTHandler.getTargetGrid().getStorageService().getInventory();
+        var playerSource = new PlayerSource(player, null);
+
+        var inventory = player.getInventory();
+        int targetSlot = inventory.getSuitableHotbarSlot();
+        var toReplace = inventory.getItem(targetSlot);
+
+        var insert = networkInventory.insert(AEItemKey.of(toReplace), toReplace.getCount(), Actionable.SIMULATE,
+                playerSource);
+        if (insert < toReplace.getCount())
+            return;
+        int targetAmount = Math.max(1, stack.getMaxStackSize() / 2);
+        var extracted = networkInventory.extract(AEItemKey.of(stack), targetAmount, Actionable.SIMULATE, playerSource);
+        if (extracted == 0)
+            return;
+
+        insert = networkInventory.insert(AEItemKey.of(toReplace), toReplace.getCount(), Actionable.MODULATE,
+                playerSource);
+        if (insert < toReplace.getCount()) {
+            toReplace.setCount(toReplace.getCount() - (int) insert);
+            inventory.setItem(targetSlot, toReplace);
+            return;
+        }
+
+        extracted = networkInventory.extract(AEItemKey.of(stack), targetAmount, Actionable.MODULATE, playerSource);
+        if (extracted == 0) {
+            inventory.setItem(targetSlot, ItemStack.EMPTY);
+            return;
+        }
+        stack.setCount((int) extracted);
+        inventory.setItem(targetSlot, stack);
+        inventory.selected = targetSlot;
+        player.connection.send(new ClientboundSetCarriedItemPacket(player.getInventory().selected));
     }
 }