+++ src/main/java/de/mari_023/ae2wtlib/AE2wtlibTags.java
@@ -1,0 +1,10 @@
+package de.mari_023.ae2wtlib;
+
+public class AE2wtlibTags {
+
+    public static final String MAGNET_MODE = "magnetMode";
+
+    public static final String PICK_BLOCK = "pick_block";
+    public static final String RESTOCK = "restock";
+    public static final String MAGNET_SETTINGS = "magnet_settings";
+}